+++ docs/.vitepress/config.ts
@@ -1,4 +1,4 @@
-import {defineConfig} from "vitepress";
+import { defineConfig} from "vitepress";
 
 export default defineConfig({
     lang: 'en-US',
@@ -6,16 +6,16 @@ export default defineConfig({
     description: 'deploy SPA so easy!',
     base: '/spa-server/',
     themeConfig: {
-        repo: 'timzaak/spa-server',
-        docsDir: 'docs',
-        docsBranch: 'master',
-        editLinks: true,
-        editLinkText: 'Edit this page on GitHub',
-        lastUpdated: 'Last Updated',
-        algolia: {
-            appId: 'NBNHWJCAL4',
-            apiKey: 'fa9bd1600dd455c6fc927a6fbafcd7b5',
-            indexName: 'spa-server',
+        lastUpdated: {
+            text: 'Last Updated'
+        },
+        search : {
+            provider: 'algolia',
+            options: {
+                appId: 'NBNHWJCAL4',
+                apiKey: 'fa9bd1600dd455c6fc927a6fbafcd7b5',
+                indexName: 'spa-server',
+            }
         },
         nav: [
             {text: 'Guide', link: '/', activeMatch: '^/$|^/guide/'},
@@ -26,7 +26,7 @@ export default defineConfig({
             },
             {
                 text: 'Release Notes',
-                link: 'https://github.com/timzaak/spa-server/releases'
+                link: 'https://github.com/fornetcode/spa-server/releases'
             }
         ],
         sidebar: {
@@ -41,14 +41,15 @@ function getGuideSidebar() {
     return [
         {
             text: 'Introduction',
-            children: [
+            items: [
                 {text: 'What is SPA-Server?', link: '/'},
                 {text: 'Getting Started', link: '/guide/getting-started'},
                 {text: 'Break Changes', link: '/guide/break-changes'}
             ],
         },
         {
-            text: 'SPA-Server', children: [
+            text: 'SPA-Server',
+            items: [
                 {text: 'Configuration', link: '/guide/spa-server-configuration'},
                 {text: 'Http API', link: '/guide/spa-server-api'},
                 {text: 'Package', link: '/guide/spa-server-release-package'},
@@ -56,25 +57,19 @@ function getGuideSidebar() {
         },
         {
             text: 'SPA-Client',
-            children: [
-                {text: 'Configuration', link: '/guide/spa-client-configuration'},
+            items: [
+                {text: 'NPM Package', link: '/guide/spa-client-npm-package'},
                 {text: 'Command Line', link: '/guide/spa-client-command-line'},
-                {text: 'NPM Package', link: '/guide/spa-client-npm-package'}
             ]
-        }, {
-            text: "Advanced",
-            children: [
-                {text: 'Uploading File Process', link: '/guide/uploading-file-process'}
-            ]
-        }
+        },
     ]
 }
 
-function getDevelopSideBar() {
+function getDevelopSideBar(){
     return [
         {
             text: 'Develop Log',
-            children: [
+            items: [
                 {text: 'Change Log', link: '/develop/change-log'},
                 {text: 'Develop Tip', link: '/develop/develop-tips'},
                 {text: 'Roadmap', link: '/develop/roadmap'},