+++ jsclient/__test__/index.spec.ts
@@ -1,7 +1,0 @@
-import test from 'ava'
-
-import { forTest } from '../index'
-
-test('sync function from native code', (t) => {
-  t.is(forTest(2), 34)
-})