+++ jsclient/run-spa.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const {runJs} = require("./index")
-
-runJs()