+++ jsclient/test/command_test_run.ts
@@ -1,0 +1,2 @@
+import command from '../src/command'
+command()