+++ jsclient/test/setEnvVars.js
@@ -1,0 +1,5 @@
+process.env.SPA_SERVER_ADDRESS='http://127.0.0.1:9000'
+
+process.env.SPA_SERVER_AUTH_TOKEN='token'
+
+process.env.SPA_UPLOAD_PARALLEL=3